+++ Remnants/Behaviours/RegisterBodySuitsBehaviour.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Remnants.utilities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -17,7 +18,7 @@ namespace Remnants.Behaviours
         private List<int> _suitsIndexList = new List<int>();
         public List<int> SuitsIndexList
         {
-            get {return _suitsIndexList;} 
+            get { return _suitsIndexList; }
         }
         #endregion
 
@@ -43,22 +44,37 @@ namespace Remnants.Behaviours
 
             _isRegistering = true;
             mls.LogInfo("Registering suits indexes.");
+            //Here I should get the list of suits that are saved.
+            //Then, I should add an if to see if the name is in the list and If it it check if it is banned or not
+            List<SuitData> suitsDataList = Remnants.Instance.RemnantsConfig.GetSuitsList();
             for (int i = 0; i < startOfRound.unlockablesList.unlockables.Count; i++)
             {
                 if (startOfRound.unlockablesList.unlockables[i].suitMaterial != null && !_suitsIndexList.Contains(i))
                 {
-                    mls.LogInfo("Register suit index of: " + startOfRound.unlockablesList.unlockables[i].suitMaterial.name);
-                    _suitsIndexList.Add(i);
+                    string suitName = startOfRound.unlockablesList.unlockables[i].suitMaterial.name;
+                    int suitDataIndex = suitsDataList.FindIndex(suitData => suitData.SuitName == suitName);
+                    if (suitDataIndex == -1)
+                    {
+                        _suitsIndexList.Add(i);
+                        suitsDataList.Add(new SuitData() { SuitName = suitName, UseSuit = true });
+                        mls.LogInfo("Register suit index of: " + suitName);
+                    }
+                    else if (suitsDataList[suitDataIndex].UseSuit)
+                    {
+                        _suitsIndexList.Add(i);
+                        mls.LogInfo("Register suit index of: " + suitName);
+                    }
                 }
             }
             mls.LogInfo("Suits indexes registered.");
             SceneManager.sceneLoaded -= RegisterSuitsIndexData;
+            Remnants.Instance.RemnantsConfig.SetSuitsList(suitsDataList);
             _isRegistering = false;
         }
 
         public void RegisterNewSuitsIndexData()
         {
-            RegisterSuitsIndexData(SceneManager.GetActiveScene(),LoadSceneMode.Additive);
+            RegisterSuitsIndexData(SceneManager.GetActiveScene(), LoadSceneMode.Additive);
         }
         #endregion
     }